Fix stale float precision in reincorporation; regenerate physics baseline
make tests summary (full suite) caught two things the earlier targeted test
runs missed after widening core Mvir/Rvir/Vvir to double:

- models/sage16/modules/sage_reincorporation/sage_reincorporation.c: Vvir and
  Rvir were copied into local `float` variables before use, silently
  re-narrowing the now-double central_halo->Vvir/Rvir - including in the
  Vvir <= Vcrit threshold branch. Widened both locals to double.
- models/sage16/modules/sage_initialise_merger_clock/_tests/test_integration_sage_initialise_merger_clock.py:
  test_output_property_types hardcoded float32 for Mvir/infallMvir; updated
  to float64 to match the new schema.

Regenerated the SAGE full-physics scientific baseline
(models/sage16/modules/_tests/baseline/physics-binary/), which is separate
from the physics-free baseline already regenerated for the core precision
change. Investigated before regenerating: per-halo differences (StellarMass,
SupernovaOutflowRate, deltaMvir, etc., 2-12% of 4196 halos) are a legitimate
cascade from the ~90 halos whose Rvir/Vvir the core fix corrected, through
threshold-sensitive physics (reincorporation's Vcrit branch, cooling timing,
AGN heating) - confirmed by aggregate mass/energy conservation (StellarMass,
ColdGas, HotGas, BlackHoleMass, Cooling, Heating totals all agree to within
0.01-0.22%, i.e. redistribution/timing shift, not new growth or corruption),
not by any additional bug beyond the reincorporation fix above.

`make tests summary` now passes cleanly (unit, integration, scientific).

+++ models/sage16/modules/sage_reincorporation/sage_reincorporation.c
@@ -73,8 +73,8 @@ int sage_reincorporation_process(struct ModuleContext *ctx, struct Halo *halos, 
     return 0;
   }
 
-  const float Vvir = central_halo->Vvir;
-  const float Rvir = central_halo->Rvir;
+  const double Vvir = central_halo->Vvir;
+  const double Rvir = central_halo->Rvir;
 
   // SN velocity 630 km/s → critical velocity = 630/sqrt(2) = 445.48 km/s
   const double Vcrit = 445.48 * REINCORPORATION_FACTOR;
